Move movie component to components folder inside movieScreen folder
