+++ src/pages/movieScreen/components/movie.js
